Feature: More companies

+++ src/window_func.h
@@ -47,7 +47,7 @@ void ShowVitalWindows();
 
 void ReInitAllWindows();
 
-void SetWindowWidgetDirty(WindowClass cls, WindowNumber number, byte widget_index);
+void SetWindowWidgetDirty(WindowClass cls, WindowNumber number, uint widget_index);
 void SetWindowDirty(WindowClass cls, WindowNumber number);
 void SetWindowClassesDirty(WindowClass cls);
